Cleanup, renaming "BCM" to "offset", moving EventWatcher to separate class.
